imports: support for ESM projects (#4)
These changes allow this package to work in an ESM project by adding
support for `.cjs` and `.cts` files when configuring the imports rules.

`Error [ERR_REQUIRE_ESM]: Must use import to load ES Module` error
happens whenever this package tries to load a `.imports.ts` file.

Because it uses `require` to load the files and this is happening in
functions that cannot be async, we can't use `await import`. Eventually
it would be good to rewrite more extensively so that `.ts` and `.js`
files can be imported directly when this is used in an ESM project, but
for now it's simple and effective to add support for `.cjs` and `.cts`
extensions, which resolves the error.