feat: switch from eslint-plugin-node to eslint-plugin-n (#8)
## Problem

`eslint-plugin-node` has been
[abandoned](https://github.com/mysticatea/eslint-plugin-node/issues/300)
for some time with no recommended replacement.

## Solution

[eslint-plugin-n](https://github.com/eslint-community/eslint-plugin-n)
is a fork that currently seems to be maintained pretty well. This
updates our dependencies and config to use `eslint-plugin-n`.

## Breaking Changes

Although it's basically a drop-in replacement, this will cause some
breaking changes so it should be in a major release.
- All config and eslint-disable comments will need to be updated from
`node/` to `n/`
- In my testing, this resulted in many new (true) errors because
`eslint-plugin-n` has been maintained and received fixes while
`eslint-plugin-node` has not. So updating will require fixing or
ignoring these new errors.