Change rules to prefer top-level `type` markers (#9)
To be able to leverage `verbatimModuleSyntax` without errors or
workarounds, we need to switch to top-level `type` markers instead of
inline markers. See examples
[here](https://github.com/import-js/eslint-plugin-import/blob/HEAD/docs/rules/consistent-type-specifier-style.md#examples).

This will ensure that TypeScript does not try to load modules from which
only types are imported, which will make it faster during build-time.
It's also a prerequisite to running TypeScript with Node directly, which
we may or may not do but would like the option to try.

Notes:
- It sounds like we should eventually remove the
`@typescript-eslint/consistent-type-imports` rule entirely based on
[this
recommendation](https://typescript-eslint.io/rules/consistent-type-imports/#comparison-with-importsnotusedasvalues--verbatimmodulesyntax).
But I thought it best to leave it there until we complete the transition
in the `platform` repo, or until we notice it causing conflicts.
- My preferred rule would be "top-level if only type imports, otherwise
inline" but I did not find any such rule. If others care deeply about
this, we could attempt to create it.

This would be released as a major/breaking change.